+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Layers](https://docs.rs/picoserve/latest/picoserve/routing/trait.Layer.html) can take ownership of requests, allowing them to:
   - Route requests to a different Router
   - Not call the next layer, but return a response
+- Implemented [Chunked](https://docs.rs/picoserve/latest/picoserve/response/chunked/struct.ChunkedResponse.html) Transfer Encoding
 
 ## [0.11.1] - 2024-06-06